Utils: Vector support for drawable reflection backend
Try to load as a vector first. If drawable is null,
try to load as a png and handle any fail as we would
before. This approach works for what we use the library
for, small lists and "one drawable at a time" types of
operations. We probably want a different approach for
speed-sensitive operations.

Change-Id: I58a10617414a1a981ff3d9d73a3cbc69b843050d