Actions: Fix default action label showing as the string resource name
Now that system action labels are translatable and stored in res,
we can no longer use a empty constructor to create ActionConfig
or ButtonConfig. The problem is when loading default configurations,
we only know the name of the label resource. So all ActionConfig
and ButtonConfig constructors require a context now. Also mandate
a context when internally creating these objects from delimited
string.

Change-Id: I7e94ebf25bc8d68a7eb5fe7be9f46cfb0d594033