ActionHandler: don't kill googlequicksearchbox or live wallpaper
Killing com.google.android.googlequicksearchbox takes down the
entire system, so don't kill it.

For some reason, when LiveWallpaper is running, it is showing as
the front process when in theory, the launcher process should be
on top. So now we have to check that the top process isn't a
LiveWallpaper service.

Change-Id: I5206e22b3b94b01ec62f8001c233441ed97d54a3